Move most of the recorder related code to separate module
